backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Shawaal Nadeem <110089331+Shawaal-Nadeem@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
google-auth = "^2.36.0"
google-auth-oauthlib = "^1.2.1"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
groq = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
